==> 01-poblacion_2011.ipynb <==
# Población por departamento según Censo 2011

Calculamos la población por depto según el Censo INE 2011 y la población con 5 años o más.
Las misma son usadas en funciones que recuperan la población desde una lista.

import pandas as pd
import os
import numpy as np

# funciones desarrolladas
from functions.agrupar_dfs_censo import *
from functions.cargar_data import *

# Carga Censo 2011
censo = cargar_censo_vars(['DPTO', 'LOC', 'PERNA01'])

# calcula pob por depto
pob = list(censo.groupby('DPTO').size())
pob

censo_plus5 = censo.loc[censo.PERNA01 >= 5]
pob5 = list(censo_plus5.groupby('DPTO').size())
pob5

np.array(pob) - np.array(pob5)
