==> 01-py_data_wrangling.ipynb <==
## data wrangling
Carga censo y filtra solo migrantes internos, luego guarda comprimido

import pandas as pd
import os
import numpy as np

# funciones desarrolladas
from functions.agrupar_dfs_censo import *
from functions.cargar_data import *

# carga censo 2011
censo = cargar_censo()

p_migr = filter_df_censo(censo)
p_migr.shape[0]

p_migr.head()

# guarda
p_migr.to_csv('tablas/censo_2011_migrantes_internos.gz',
               index=False,
               compression="gzip")
