==> 02-eda_mapas.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.font_manager as font_manager

# funciones desarrolladas
from functions.cargar_data import *
from functions.guardar_data import *
from functions.graficas import *

variables = ['DPTO', 'LOC','PERMI07', 'PERMI07_1', 'PERMI07_2', 'PERMI07_3', 'PERMI07_4']

censo = cargar_censo_vars(variables)

censo.head()

def conteo_migrantes(df):
    "Conteo de migrantes interno por departamento"
    lista_vars = ['total', 'cat_a', 'cat_b', 'cat_c']
    df[lista_vars[0]] = 1
    [df.insert(len(df.columns), lista_vars[i], 0) for i in range(1,4)]
    # destino Mvdeo.
    df.loc[(df.PERMI07 == 3) & (df.DPTO==1), lista_vars[1]] = 1
    # destino interior
    df.loc[(df.PERMI07 == 3) & (df.PERMI07_2== 1), lista_vars[2]] = 1
    # entre interior
    df.loc[(df.PERMI07 == 3) & (df.PERMI07_2!=1) & (df.DPTO!=1) & (df.PERMI07_2.isnull()==False), lista_vars[3]] = 1

    # agrupa
    df_group = df.groupby('DPTO').agg(
                {   'total': sum,
                     lista_vars[1]: sum,
                     lista_vars[2]: sum,
                     lista_vars[3]: sum
                })

    df_group['total_mig'] = df_group[lista_vars[1]] + df_group[lista_vars[2]] + df_group[lista_vars[3]]
    
    return df_group

migrantes = conteo_migrantes(censo)
migrantes

# Exporta CSV
# migrantes.to_csv('tablas/migrantes_por_depto.csv')

censo.groupby(['PERMI07_2', 'PERMI07_3']).DPTO.count().reset_index()

def conteo_migrantes_por_loc(df):
    "Conteo por localidad de migrantes internos con origen en Mvdeo."
    df_filtered = df.loc[(df.PERMI07 == 3) & (df.PERMI07_2== 1)].copy()
    df_filtered['CODLOC'] = df_filtered.DPTO.astype(str) + pad(df_filtered, 'LOC', 3)
    df_grouped = df_filtered.groupby(['CODLOC']).DPTO.count().reset_index()
    df_grouped.rename({'CODLOC':'loc_destino', 'DPTO':'personas'}, inplace=True, axis=1)
    
    return df_grouped.sort_values('personas', ascending=False)

mig_loc = conteo_migrantes_por_loc(censo)

# exporta
# mig_loc.to_csv('tablas/migrantes_no_mvo_por_loc.csv', index=False)
# campos = 'Integer','Integer'
# export_csvt(campos, 'tablas/migrantes_no_mvo_por_loc.csvt')

merge_loc = loc_decode(mig_loc)
merge_loc.head()

# porcentaje de personas migrantes internas corresppondientes a las 7
# localidades que más reciben migrantes
print(round(merge_loc.head(7).personas.sum() / merge_loc.personas.sum() * 100, 2 ))

# identifica casos de AM
am = pd.read_csv('tablas/localidades_am_mvo.csv', usecols=['codloc'], dtype= {'codloc': str,})
migrantes_mvo = merge_loc.merge(am, how='left', left_on='loc_destino', right_on='codloc')
migrantes_mvo.loc[~migrantes_mvo.codloc.isnull(), 'categoria'] = 'AREA METRO'
migrantes_mvo.loc[migrantes_mvo.codloc.isnull(), 'categoria'] = migrantes_mvo.departamento

# calcula n y porcentajes
n_mig = migrantes_mvo.groupby(['categoria']).personas.sum()
percent_mig = (migrantes_mvo.groupby(['categoria']).personas.sum()/migrantes_mvo.personas.sum()*100).round(2).rename('porcentaje', axis=1)
migrantes_mvo_am= pd.concat([n_mig, percent_mig], axis=1).reset_index()

# agrega total
total = pd.DataFrame([['Total', sum(migrantes_mvo_am.personas), sum(migrantes_mvo_am.porcentaje)]], columns=['categoria', 'personas', 'porcentaje'])
migrantes_mvo_am = migrantes_mvo_am.append(total)

migrantes_mvo_am.loc[:, 'categoria'] = format_depto(migrantes_mvo_am, 'categoria')

migrantes_mvo_am.rename({'categoria':'Entidad geográfica'}, inplace=True, axis=1)

print(migrantes_mvo_am.to_markdown(index=False, tablefmt='pretty'))
