==> 03-dd_vecindad.ipynb <==
# Calcula vecindad entre departamentos

Se calcula la vecindad y el largo de la frontera compartida entre departamentos

import pandas as pd
import fiona
from shapely.geometry import shape, mapping
import itertools
from collections import OrderedDict

# carga la capa como un iterable
polys = fiona.open("capas/ine_deptos_generalizada.gpkg")

def shared_boundary(file):
    "Calculo de vecindad y largo de fronteras"
    polys = fiona.open(file)
    shared = []

    for pol1, pol2 in itertools.combinations(polys, 2):
        geompol1 = shape(pol1['geometry'])
        geompol2 = shape(pol2['geometry'])
        # Si se tocan, calcula largo
        if geompol1.touches(geompol2):
            length = shape(mapping(geompol1.intersection(geompol2))).length
            shared.append([tuple(pol1['properties'].items())[0], tuple(pol2['properties'].items())[0], int(length)])
    
    # recupera valores mediante list comprehension
    shared = list(zip([x[0][1] for x in shared],
                  [x[1][1] for x in shared],
                  [x[2] for x in shared]                  
                 ))
    # cambia a DF
    df = pd.DataFrame(shared, columns=['poly1', 'poly2', 'length']).sort_values(by=['poly1', 'poly2'])
    # elimina los registro que se tocan en un sólo punto (largo=0)
    df = df.loc[df.length > 0].reset_index(drop=True)

    return df

# ruta a la capa de departamentos
file="capas/ine_deptos_generalizada.gpkg"

# aplica función
vecinos = shared_boundary(file)

# renombra variables
vecinos.rename({'poly1': 'dep1', 'poly2': 'dep2', 'length': 'largo_limite'}, inplace=True, axis=1)
print(vecinos.shape)
vecinos.head()

def codificar(df, var1, var2):
    "Función para codificar"
    df_coded = df.copy()
    df_coded['cod'] = (df_coded[var1].astype(str) + df_coded[var2].astype(str).str.zfill(2)).astype(int)
    
    return df_coded

# codifica
df1 = codificar(vecinos, 'dep1', 'dep2')
df2 = codificar(vecinos, 'dep2', 'dep1')
df2.columns = ['dep2', 'dep1', 'largo_limite', 'cod']

# hace un append de ambos DF
deptos_vecinos = df1.append(df2).sort_values(['dep1', 'dep2']).reset_index(drop=True)

print(deptos_vecinos.shape)
deptos_vecinos.head()

# exporta
deptos_vecinos.to_csv('tablas/deptos_vecinos.csv', index=False)
