==> 04-dd_procesamiento.ipynb <==
import pandas as pd
import numpy as np

# funciones desarrolladas
from functions.agrupar_dfs_censo import *
from functions.cargar_data import *
from functions.impresion import *

# carga datos diádicos
pbi = cargar_pbi()
md = cargar_matriz_distancias()
vecindad = cargar_vecindad()
dist_loc = pd.read_csv('tablas/df_distancias_localidades.csv', usecols=['cod','distancia_m'])

# carga capas de datos geográficos
deptos, deptos_sim, localidad, centro_pobl, capital = cargar_datos_geo()

# carga migrantes internos
p_migr = cargar_migrantes_internos()

# carga poblacion por depto
pobl = recuperar_poblacion_2011()

# menores de 25 años por depto
censo_edad = cargar_censo_vars(['DPTO', 'PERNA01'])
censo_edad = censo_edad.loc[censo_edad.PERNA01 < 25].reset_index(drop=True)
menores_25 = censo_edad.groupby('DPTO').size()
menores_25.name = 'menores_25'

# carga cantidad de empresas industriales por depto (OTU, OPP, 2013)
# Fuente: Núcleo Interdisciplinario de Estudios de Desarrollo Territorial (NIEDT)
# en base a INE (Directorio de Empresas y Establecimientos) y relevamiento propio
# Base: Total de empresas industriales de más de 5 empleados
empresas = pd.read_csv('tablas/empresas_por_depto.csv', sep=';', usecols=['DPTO', 'empresas'])

# genera código de díada
p_migr['cod'] = (p_migr['loc_origen'].astype(str) + p_migr['loc_destino'].astype(str)).astype(int)

# pega distancias entre localidades
p_migr = p_migr.merge(dist_loc, how='left', on='cod')
p_migr.head(2)

print(p_migr.shape)
print(p_migr.columns)
p_migr.head()

p_migr.loc[(p_migr.depto_origen==1) & (p_migr.depto_destino==3), 'distancia_m'].max()

p_migr.loc[(p_migr.depto_origen==3) & (p_migr.depto_destino==1), 'distancia_m'].max()

# genera un sólo dataframe solo para deptos
flujos_deptos = p_migr[['depto_origen', 'depto_destino', 'PERNA01']].copy()

flujos_deptos = flujos_deptos[flujos_deptos.depto_origen != flujos_deptos.depto_destino]

flujos_deptos['personas_mig'] = 1
flujos_deptos['personas_mig_18_25'] = 0
flujos_deptos['personas_mig_26_35'] = 0

flujos_deptos.loc[flujos_deptos.PERNA01.between(18, 25), 'personas_mig_18_25'] = 1
flujos_deptos.loc[flujos_deptos.PERNA01.between(26, 35), 'personas_mig_26_35'] = 1

flujos_deptos.drop('PERNA01', axis=1, inplace=True)


# agrupa y cuenta
grupo = flujos_deptos.groupby(by=['depto_origen', 'depto_destino']).sum()

print('Cantidad de díadas ', grupo.shape[0])

grupo.head(3)

data_webmap = grupo.reset_index().loc[:, ['depto_origen', 'depto_destino', 'personas_mig']]
data_webmap.to_csv('webmap_flows/flows.csv', index=False)

# flujos en movimientos mayores a 100km
flujos_deptos_100 = p_migr[['depto_origen', 'depto_destino', 'distancia_m']].copy()

flujos_deptos_100 = flujos_deptos_100[flujos_deptos_100.depto_origen != flujos_deptos_100.depto_destino]

flujos_deptos_100['personas_mig'] = 1

# agrupa y cuenta
grupo_100 = pd.DataFrame(flujos_deptos_100.loc[flujos_deptos_100.distancia_m > 100000].groupby(by=['depto_origen', 'depto_destino']).sum())

print('Cantidad de díadas ', grupo_100.shape[0])

grupo_100.head(3)

# genera tabla pivot con los flujos de departamento a departamento
matrix = pd.pivot_table(flujos_deptos,
                        index ='depto_origen',
                        columns='depto_destino',
                        fill_value=0,
                        aggfunc=sum,
                        margins=True,
                        margins_name='Total')

# # guarda en csv
# matrix = pd.pivot_table(flujos_deptos,
#                         index ='depto_origen',
#                         columns='depto_destino',
#                         fill_value=0,
#                         aggfunc=sum,
#                         margins=False)

# matrix.to_csv('tablas/matriz_deptos.csv', header=True)

# # guarda en latex
# # lista de nombres deptos para generar tabla en latex
# nomdep = ['Mvdeo.', 'Artigas', 'Can.', 'C. Largo', 'Colonia', 'Durazno',
#           'Flores', 'Florida', 'Lavalleja','Maldonado', 'Paysandú', 'R. Negro', 'Rivera',
#           'Rocha', 'Salto', 'San José', 'Soriano', 'Tacuarembó', 'T. y Tres']

# coddep = np.arange(1, 20, 1).tolist()

# # guarda tabla en latex
# # cabecera = datos_dpto.sort_values('DPTO')['NOMBRE'].to_list()
# # cabecera.append('Total')

# matrix_tex = matrix
# matrix_tex.rename(index=dict(zip(coddep, nomdep)), inplace=True)

# # setea ancho de columnas
# ancho = 'p{0.7cm}'
# colformato='l' + ancho * 20

# matrix_tex.to_latex(buf= "tablas/matriz_orig_dest.tex", bold_rows=False,
#                     column_format = colformato,
#                     caption= 'Matriz de movimientos entre departamentos (Censo INE 2011).')

# unimos todo en un dataframe de díadas

# migrantes
df_agrupado = grupo.reset_index()

# agrega codigo único
df_agrupado.insert(0, 'cod', (df_agrupado['depto_origen'].astype(str)
                   + df_agrupado['depto_destino'].astype(str).str.zfill(2)).astype(int))

df_agrupado.head(2)

df_agrupado_100k = grupo_100.reset_index()
# agrega codigo único
df_agrupado_100k.insert(0, 'cod', (df_agrupado_100k['depto_origen'].astype(str)
                   + df_agrupado_100k['depto_destino'].astype(str).str.zfill(2)).astype(int))

df_agrupado_100k.drop(['depto_origen', 'depto_destino', 'distancia_m'], axis=1, inplace=True)
df_agrupado_100k.rename({'personas_mig': 'personas_mig_100k'}, axis=1, inplace=True)
df_agrupado_100k.head(2)

# pega
df_agrupado = df_agrupado.merge(df_agrupado_100k, how='left', on='cod')

df_agrupado['personas_mig_100k'] = df_agrupado['personas_mig_100k'].fillna(0).astype(int)

df_agrupado.head(2)

# pega masas (población!)
df_agrupado = df_agrupado.merge(pobl, how='left', left_on='depto_origen', right_on='DPTO')

df_agrupado = df_agrupado.merge(pobl, how='left', left_on='depto_destino', right_on='DPTO')

def drop_deptos(df):
    df = df.drop(['DPTO_x', 'DPTO_y'], axis=1, inplace=True)

drop_deptos(df_agrupado)

df_agrupado.head(2)

# pega matriz de distancias
df_agrupado = df_agrupado.merge(md, how='left', on='cod')
df_agrupado = df_agrupado.drop(['cod_ori', 'cod_des'], axis=1)

# convierte distancia a KM
df_agrupado['dist_km'] = round(df_agrupado.distancia / 1000).astype(int)

df_agrupado.head(2)

# pega datos del PBI
df_agrupado = df_agrupado.merge(pbi, how='left', left_on='depto_origen',  right_on='DPTO')
df_agrupado = df_agrupado.merge(pbi, how='left', left_on='depto_destino', right_on='DPTO')

# convierte unidades
df_agrupado['pbi_origen_millardos']  = round(df_agrupado.miles_de_pesos_x / 1000000, 3)
df_agrupado['pbi_destino_millardos'] = round(df_agrupado.miles_de_pesos_y / 1000000, 3)

drop_deptos(df_agrupado)

df_agrupado.head(2)

# pega vecindad
df_agrupado = df_agrupado.merge(vecindad, how='left', on='cod').drop(['dep1', 'dep2'], axis=1)
df_agrupado['dummy_limit'] = True
df_agrupado.loc[df_agrupado.largo_limite.isnull(), 'dummy_limit'] = False

df_agrupado.head(2)

# pega empresas
df_agrupado = df_agrupado.merge(empresas, how='left', left_on='depto_origen', right_on='DPTO')
df_agrupado = df_agrupado.merge(empresas, how='left', left_on='depto_destino', right_on='DPTO')

drop_deptos(df_agrupado)

df_agrupado.head(2)

# pega menores de 25 años
df_agrupado = df_agrupado.merge(menores_25, how='left', left_on='depto_origen', right_on='DPTO')
df_agrupado = df_agrupado.merge(menores_25, how='left', left_on='depto_destino', right_on='DPTO')
df_agrupado.head(2)

dict_rename = {
    'miles_de_pesos_x':'pbi_origen',
    'miles_de_pesos_y':'pbi_destino',
    'porcentaje_pbi_x': 'pbi_porcen_ori',
    'porcentaje_pbi_y': 'pbi_porcen_des',
    'empresas_x':'empresas_origen',
    'empresas_y': 'empresas_destino',
    'miles_de_pesos':'pbi_destino',
    'distancia': 'dist',
    'poblacion_x': 'pob_origen',
    'poblacion_y': 'pob_destino',
    'menores_25_x': 'menores_25_origen',
    'menores_25_y': 'menores_25_destino'
}

df_agrupado.rename(dict_rename, inplace=True, axis=1)

df_agrupado.head(2)

# genera otro DF igual a df_agrupado, pero llamado dd_deptos
dd_deptos = df_agrupado

print(dd_deptos.shape)

## pob destino y origen a miles
dd_deptos['pob_destino_k'] = dd_deptos['pob_destino']/1000
dd_deptos['pob_origen_k'] = dd_deptos['pob_origen']/1000

# decodifica el departamento (códigos INE)
dd_deptos['nom_depto_orig'] = decode_depto(dd_deptos, 'depto_origen')
dd_deptos['nom_depto_des'] = decode_depto(dd_deptos, 'depto_destino')

dd_deptos.head(2)

censo = cargar_censo_nrows(2)
censo.columns

# agrega edad promedio por departamento
cols = ['DPTO', 'PERNA01', 'POBPCOAC']
censo = cargar_censo_vars(cols)
censo.head()

emedia = censo.groupby('DPTO').mean()['PERNA01'].to_frame()
emedia.head(2)

emedia = censo.groupby('DPTO').mean()['PERNA01'].to_frame()
emedia.head(2)

ocupados = censo.loc[censo.POBPCOAC==2, 'DPTO'].to_frame()
ocup_gr = ocupados.groupby('DPTO').size()

pet = censo.loc[(censo.PERNA01 >= 15) & (censo.PERNA01 < 65), 'DPTO'].to_frame()
pet_gr = pet.groupby('DPTO').size().values

porc_ocupados = (ocup_gr / pet_gr)*100

emedia['porc_ocupados'] = porc_ocupados

dd_deptos_new = dd_deptos.merge(emedia, left_on='depto_origen', right_on='DPTO')
dd_deptos_new = dd_deptos_new.merge(emedia, left_on='depto_destino', right_on='DPTO')

dict_rename = {'PERNA01_x': 'edad_prom_ori',
               'porc_ocupados_x': 'porc_ocupados_ori',
               'PERNA01_y': 'edad_prom_des',
               'porc_ocupados_y': 'porc_ocupados_des'}

dd_deptos_new.rename(dict_rename, inplace=True, axis=1)

dd_deptos_new.head(2)

#### Tasa de desempleo 2011
empleo = pd.read_csv('tablas/tasa-desempleo.csv', sep=';', decimal=',')

emp_m = empleo.loc[(empleo.Sexo=='Mujeres') & (empleo.Departamento!='Total'), '2011']
emp_v = empleo.loc[(empleo.Sexo=='Varones') & (empleo.Departamento!='Total'), '2011']
emp_mv = empleo.loc[(empleo.Sexo=='Total') & (empleo.Departamento!='Total'), '2011']
dep = empleo.loc[(empleo.Sexo=='Total') & (empleo.Departamento!='Total'), 'Departamento']
emp = [encode_depto_pretty(dep), emp_m, emp_v, emp_mv]
[i.reset_index(inplace=True, drop=True) for i in emp]

emp_concat = pd.concat(emp, axis=1)

emp_concat.head()


emp_concat.columns = [
    'depto',
    'tasa_desemp_mujeres_2011_ori',
    'tasa_desemp_varones_2011_ori',
    'tasa_desemp_total_2011_ori'
]


dd_deptos_new = dd_deptos_new.merge(
    emp_concat, left_on='depto_origen', right_on='depto')

emp_concat.columns = ['depto',
                      'tasa_desemp_mujeres_2011_des',
                      'tasa_desemp_varones_2011_des',
                      'tasa_desemp_total_2011_des'
                      ]

dd_deptos_new = dd_deptos_new.merge(
    emp_concat, left_on='depto_destino', right_on='depto')
cols = ['depto_x', 'depto_y']
dd_deptos_new.drop(cols, axis=1, inplace=True)

dd_deptos_new.head()


# calcula logaritmos
def logaritmos(df, variables):
    "Calcula logaritmos para N variables"
    new_vars = ['log_' + i for i in variables]
    n = len(variables)
    for i in range(n):
        df[new_vars[i]] = np.log(df[variables[i]])

    return df

# definir lista de variables sobre las cuales se quiere el logaritmo
variables = [
    'pbi_destino',
    'dist_km',
    'pbi_origen_millardos',
    'pbi_destino_millardos',
    'pob_destino_k',
    'pob_origen_k',
    'edad_prom_des',
    'porc_ocupados_des'
]

dd_deptos_new = logaritmos(dd_deptos_new, variables)

dd_deptos_new.head(2)

# carga empresas
pymes = pd.read_csv('tablas/mypymes.csv')
pymes.head()

# guarda
dd_deptos_new.to_csv('tablas/dd_deptos.csv', index=False, sep=';', decimal=',')
