==> 05-dd_localidades.ipynb <==
import pandas as pd
import numpy as np
import itertools

# funciones desarrolladas
from functions.cargar_data import *

# calcula migrantes internos entre localidades
pmigr = cargar_migrantes_internos()

mig_loc  = pmigr.groupby(['loc_origen','loc_destino']).size().to_frame('personas_mig').reset_index()
mig_loc['cod'] = (mig_loc.loc_origen.astype(str) + mig_loc.loc_destino.astype(str)).astype(int)

print(mig_loc.shape)
mig_loc.head(2)

# levanta población por localidad
poblacion = pd.read_csv('tablas/localidades_censales_2011.csv', usecols=['codloc','poblacion'])

print(poblacion.shape)
poblacion.head(2)

# levanta distancias
dist = pd.read_csv('tablas/df_distancias_localidades.csv', usecols=['cod','distancia_m'])
print(dist.shape)
dist.head(2)

def permutations(array):
    "Genera las combinaciones de pares de códigos posibles devolviendo un DF"
    _combi = list(itertools.permutations(array, 2))
    _cod_ori = [x[0] for x in _combi]
    _cod_des = [x[1] for x in _combi]
    _combi_codes = [int(str(x[0]) + str(x[1])) for x in _combi]
    data_tuples = list(zip(_combi_codes, _cod_ori, _cod_des, ))
    cod = pd.DataFrame(data_tuples, columns=['cod', 'cod_ori', 'cod_des'])
    return cod

cod = permutations(poblacion.codloc.values)
cod.head(2)

# pega población en origen y destino
dd = cod.merge(poblacion, left_on='cod_ori', right_on='codloc', how='left')
dd = dd.merge(poblacion, left_on='cod_des', right_on='codloc', how='left')
dd.drop(['codloc_x', 'codloc_y'], axis=1, inplace=True)
dd.rename({'poblacion_x': 'poblacion_ori', 'poblacion_y': 'poblacion_des'}, inplace=True, axis=1)
print(dd.shape)
dd.head()

# pega población migrante por díada
dd = dd.merge(mig_loc, on='cod', how='left').drop(['loc_origen', 'loc_destino'], axis=1)
dd.head()

# pega distancias
dd = dd.merge(dist, on='cod')

dd.head()

# exporta a CSV
dd.to_csv('tablas/dd_localidades.csv', index=False)
