==> 06-MIE_localidades.ipynb <==
import pandas as pd
import geopandas as gpd
import os
import matplotlib.pyplot as plt
import pylab as pl
import numpy as np
import statsmodels.api as sm
import statsmodels.formula.api as smf
import tabulate
from operator import add

from shapely import wkt
from shapely.geometry import shape, LineString, Point
    
# funciones desarrolladas
from functions.agrupar_dfs_censo import *
from functions.cargar_data import *
from functions.impresion import *

pmigr = cargar_migrantes_internos()

pmigr.head()

mig_loc  = pmigr.groupby(['loc_origen','loc_destino']).size().to_frame('personas_mig').reset_index()

poblacion = pd.read_csv('tablas/localidades_censales_2011.csv', usecols=['codloc','poblacion'])

poblacion.head()

dd_loc = mig_loc.merge(poblacion, left_on='loc_origen', right_on='codloc')
dd_loc = dd_loc.merge(poblacion, left_on='loc_destino', right_on='codloc')

dd_loc.drop(['codloc_x', 'codloc_y'], inplace=True, axis=1)
dd_loc.rename({'poblacion_x':'pobl_origen', 'poblacion_y':'pobl_destino'}, inplace=True, axis=1)

dd_loc.head()

dd_loc.to_csv('tablas/dd_localidades.csv', index=False)
